feat(play): /profile 분리 + 헤더 프로필 아바타 메뉴
- 내 정보(닉네임/자기소개/링크) 편집을 /my 에서 /profile 로 분리
- 헤더 내 작품/로그아웃 버튼 제거 → 아바타 메뉴(내 정보/내 작품/로그아웃)
