+++ specs/play-igrus/spec.md
@@ -106,9 +106,11 @@ IGRUS 가입자(동아리 비회원 포함 — 인하대생 누구나)가 본인
 - `/login`: 학번(8자리)+비밀번호. 회원가입은 www 링크.
 - `/submit`: RHF 제출 폼 (설명 50자 카운터, 이미지 드래그 앤 드롭·미리보기, 필수 뱃지).
 - `/edit/{id}`: 제출 폼 재사용 — 대기 중 수정본이 있으면 그 내용을 프리필.
-- `/my`: 공개 프로필 편집(닉네임/자기소개/링크 +·🗑 — igrus `users` 테이블에 저장,
-  `PATCH /api/v1/mypage/profile` 직접 호출) + 내 제출 상태 (심사중/승인됨/반려됨+사유,
-  수정 심사중/수정 반려 표시, 수정 버튼).
+- `/profile`: 공개 프로필(내 정보) 편집 — 닉네임/자기소개/링크 +·🗑 (igrus `users`
+  테이블에 저장, `PATCH /api/v1/mypage/profile` 직접 호출).
+- `/my`: 내 제출 상태 (심사중/승인됨/반려됨+사유, 수정 심사중/수정 반려 표시, 수정 버튼).
+- 헤더: 출시하기(+운영진 검수)와 프로필 아바타(스켈레톤) — 아바타 탭 →
+  내 정보 / 내 작품 / 로그아웃 메뉴 (내 작품·로그아웃 버튼은 메뉴로 이동).
 - `/admin`: 운영진 검수 (대기/승인/반려 탭, 승인·반려 버튼).
 
 ## 배포/인프라
